# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Grado API",
        "description": "API para la gestión de trabajos de grado. Proporciona endpoints para administradores, gestores y evaluadores.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/notificaciones/{id}/leida": {
            "post": {
                "tags": [
                    "Notificaciones"
                ],
                "summary": "Marcar notificación como leída",
                "operationId": "5f7e0bff1d85938afc4a9fdbb23bcfa7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID de la notificación",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificación marcada como leída",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notificación no encontrada"
                    }
                }
            }
        },
...
    },
...
}